add many line message and file, image inline in editor
