+++ package.json
@@ -37,6 +37,7 @@
     "react-dom": "^19.1.0",
     "react-easy-crop": "^5.5.0",
     "react-i18next": "^15.4.1",
+    "react-textarea-autosize": "^8.5.9",
     "tiptap-extension-resize-image": "^1.2.2"
   },
   "devDependencies": {
